register checks and db connection for login-register

+++ src/main/java/com/example/yourdoctorsapp/UserRepository.java
@@ -1,9 +1,13 @@
 package com.example.yourdoctorsapp;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.data.repository.CrudRepository;
 
 
 
-public interface UserRepository extends JpaRepository<Patients,Long > {
+public interface UserRepository extends JpaRepository<Patients,BigDecimal> {
+    
 
 }
